ny Handlebar helper for å inkludere 3 desimaler fordi det er ønskelig å vise det slik
Fikser etter merge av ny sats

+++ formidling/src/main/java/no/nav/ung/sak/formidling/innhold/GrunnlagOgTilkjentYtelse.java
@@ -27,6 +27,6 @@ public record GrunnlagOgTilkjentYtelse(
     Integer antallBarn,
     int barnetillegg,
     long dagsatsTilkjentYtelse,
-    BigDecimal utbetalingsgrad
+    long utbetalingsgrad
 ) {
 }
